requirements.txt
joblib==1.2.0
mysql-connector-python==8.0.31
numpy==1.23.4
pandas==1.5.1
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.6
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
